Added front-end components and admin page/categories

+++ ReactBoard.Domain/Entities/Image/IImageMetadataRepository.cs
@@ -2,7 +2,7 @@
 
 namespace ReactBoard.Domain.Entities.Image
 {
-    public interface IImageMetadataRepository : IEntityRepository<ImageMetadata, int>
+    public interface IImageMetadataRepository : IEntityRepository<ImageMetadata, long>
     {
     }
 }
