+++ ReactBoard.Domain/Services/Image/ImageMetadataService.cs
@@ -3,7 +3,7 @@ using ReactBoard.Domain.Entities.Image;
 
 namespace ReactBoard.Domain.Services.Image
 {
-    public sealed class ImageMetadataService : EntityService<ImageMetadata, int>, IImageMetadataService
+    public sealed class ImageMetadataService : EntityService<ImageMetadata, long>, IImageMetadataService
     {
         public ImageMetadataService(IImageMetadataRepository repository) : base(repository) { }
     }
